feat(data): implement data context with calculations and chart integration
